% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_sharpness.R
\name{compute_sharpness_distribution}
\alias{compute_sharpness_distribution}
\title{Compute sharpness for distribution diagnostics}
\usage{
compute_sharpness_distribution(
  simulated_distributions,
  probs = seq(0.05, 0.95, by = 0.05),
  c = 1e-08
)
}
\arguments{
\item{simulated_distributions}{List of numeric vectors.}

\item{probs}{Probability grid.}

\item{c}{Stabilizing constant.}
}
\value{
A list with average quantile band width and relative width.
}
\description{
Compute sharpness for distribution diagnostics
}
