% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_state.R
\name{diagnostic_switching_rate}
\alias{diagnostic_switching_rate}
\title{Compute switching-rate diagnostics}
\usage{
diagnostic_switching_rate(observed_states, simulated_states)
}
\arguments{
\item{observed_states}{Observed or decoded state path.}

\item{simulated_states}{Matrix of simulated state paths.}
}
\value{
Diagnostic row.
}
\description{
Compute switching-rate diagnostics
}
