% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_state.R
\name{diagnostic_transition_counts}
\alias{diagnostic_transition_counts}
\title{Compute transition-count diagnostics}
\usage{
diagnostic_transition_counts(observed_states, simulated_states)
}
\arguments{
\item{observed_states}{Observed or decoded state path.}

\item{simulated_states}{Matrix of simulated state paths.}
}
\value{
Diagnostic row.
}
\description{
Compute transition-count diagnostics
}
