% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{extract_hmmssf_parameters}
\alias{extract_hmmssf_parameters}
\title{Extract HMM-SSF parameters}
\usage{
extract_hmmssf_parameters(fit)
}
\arguments{
\item{fit}{A fitted object.}
}
\value{
A list of available parameters.
}
\description{
Extract HMM-SSF parameters
}
