% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{extract_initial_distribution}
\alias{extract_initial_distribution}
\title{Extract the initial distribution}
\usage{
extract_initial_distribution(fit)
}
\arguments{
\item{fit}{A fitted object.}
}
\value{
A probability vector.
}
\description{
Extract the initial distribution
}
