% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{extract_state_likelihoods}
\alias{extract_state_likelihoods}
\title{Extract state likelihoods}
\usage{
extract_state_likelihoods(fit, required = TRUE)
}
\arguments{
\item{fit}{A fitted object.}

\item{required}{Should missing likelihoods trigger an error?}
}
\value{
Matrix of log likelihoods or \code{NULL}.
}
\description{
Extract state likelihoods
}
