% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{extract_state_specific_kernels}
\alias{extract_state_specific_kernels}
\title{Extract state-specific movement kernels}
\usage{
extract_state_specific_kernels(fit, required = TRUE)
}
\arguments{
\item{fit}{A fitted object.}

\item{required}{Should missing kernels trigger an error?}
}
\value{
List of kernels or \code{NULL}.
}
\description{
Extract state-specific movement kernels
}
