% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{extract_transition_probabilities}
\alias{extract_transition_probabilities}
\title{Extract transition probabilities}
\usage{
extract_transition_probabilities(fit)
}
\arguments{
\item{fit}{A fitted object.}
}
\value{
A transition matrix, transition array, or transition function.
}
\description{
Extract transition probabilities
}
