% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{extract_viterbi_path}
\alias{extract_viterbi_path}
\title{Extract Viterbi path}
\usage{
extract_viterbi_path(fit, required = TRUE)
}
\arguments{
\item{fit}{A fitted object.}

\item{required}{Should missing paths trigger an error?}
}
\value{
Integer state path or \code{NULL}.
}
\description{
Extract Viterbi path
}
