% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMSSFGenerativeRepair-package.R
\docType{package}
\name{HMMSSFGenerativeRepair-package}
\alias{HMMSSFGenerativeRepair}
\alias{HMMSSFGenerativeRepair-package}
\title{HMMSSFGenerativeRepair package}
\description{
Development scaffold for HMM-SSF generative validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aureliennicosiaulaval.github.io/HMMSSFGenerativeRepair/}
}

}
\author{
\strong{Maintainer}: Aurelien Nicosia \email{116816597+AurelienNicosiaULaval@users.noreply.github.com}

}
\keyword{internal}
