% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_hmmSSF.R
\name{validate_hmmssf_adapter}
\alias{validate_hmmssf_adapter}
\title{Validate an adapted HMM-SSF object}
\usage{
validate_hmmssf_adapter(x)
}
\arguments{
\item{x}{Adapted object returned by \code{\link[=as_gmov_hmmssf]{as_gmov_hmmssf()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Validate an adapted HMM-SSF object
}
